+++ src/onegov.election_day/onegov/election_day/formats/__init__.py
@@ -1,5 +1,4 @@
 from onegov.election_day.formats.election import import_election_internal
-from onegov.election_day.formats.election import import_election_sesam
 from onegov.election_day.formats.election import import_election_wabsti_majorz
 from onegov.election_day.formats.election import import_election_wabsti_proporz
 from onegov.election_day.formats.election import import_election_wabstic_majorz
@@ -14,7 +13,6 @@ from onegov.election_day.formats.vote.wabstic import import_vote_wabstic
 
 __all__ = [
     'import_election_internal',
-    'import_election_sesam',
     'import_election_wabsti_majorz',
     'import_election_wabsti_proporz',
     'import_election_wabstic_majorz',