+++ src/onegov.election_day/onegov/election_day/formats/election/__init__.py
@@ -2,8 +2,6 @@ from onegov.election_day.formats.election.internal \
     import import_election_internal
 from onegov.election_day.formats.election.parties \
     import import_party_results
-from onegov.election_day.formats.election.sesam \
-    import import_election_sesam
 from onegov.election_day.formats.election.wabsti_majorz \
     import import_election_wabsti_majorz
 from onegov.election_day.formats.election.wabsti_proporz \
@@ -16,7 +14,6 @@ from onegov.election_day.formats.election.wabstic_proporz \
 
 __all__ = [
     'import_election_internal',
-    'import_election_sesam',
     'import_election_wabsti_majorz',
     'import_election_wabsti_proporz',
     'import_election_wabstic_majorz',